services/python/analytics-pipeline/src/requirements/function.txt
# Cloud Function
google-cloud-storage==1.19.0
google-cloud-bigquery==1.15.0
